backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]==2.10.3
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
python-jose==3.3.0
bcrypt==4.1.2
python-multipart==0.0.20
email-validator==2.1.0
redis==5.2.0
APScheduler==3.10.4
